requirements.txt
Django==1.6.1
dj-database-url==0.3.0
dj-static==0.0.5
django-toolbelt==0.0.1
djangotoolbox==1.6.2
gunicorn==18.0
mongoengine==0.8.6
psycopg2==2.5.2
pymongo==2.8
pystache==0.5.3
static==1.0.2

greenlet==0.4.0
django-storages==1.1.8
boto==2.27.0